# Benchmark: use `not_highly_aligned_allocator` in more places