--- a/benchmarks/src/std_copy.cpp
+++ b/benchmarks/src/std_copy.cpp
@@ -8,14 +8,15 @@
 #include <type_traits>
 #include <vector>
 
-#include <udt.hpp>
-#include <utility.hpp>
+#include "skewed_allocator.hpp"
+#include "udt.hpp"
+#include "utility.hpp"
 
 template <typename Contained>
 void handwritten_loop(benchmark::State& state) {
     const size_t r0      = static_cast<size_t>(state.range(0));
-    const auto in_buffer = random_vector<Contained>(r0);
-    std::vector<Contained> out_buffer(r0);
+    const auto in_buffer = random_vector<Contained, not_highly_aligned_allocator>(r0);
+    std::vector<Contained, not_highly_aligned_allocator<Contained>> out_buffer(r0);
     for ([[maybe_unused]] auto _ : state) {
         benchmark::DoNotOptimize(in_buffer.data());
         const Contained* in_ptr           = in_buffer.data();
@@ -32,8 +33,8 @@ void handwritten_loop(benchmark::State& state) {
 template <typename Contained>
 void handwritten_loop_n(benchmark::State& state) {
     const size_t r0      = static_cast<size_t>(state.range(0));
-    const auto in_buffer = random_vector<Contained>(r0);
-    std::vector<Contained> out_buffer(r0);
+    const auto in_buffer = random_vector<Contained, not_highly_aligned_allocator>(r0);
+    std::vector<Contained, not_highly_aligned_allocator<Contained>> out_buffer(r0);
     for ([[maybe_unused]] auto _ : state) {
         benchmark::DoNotOptimize(in_buffer.data());
         const Contained* const in_ptr = in_buffer.data();
@@ -50,8 +51,8 @@ template <typename Contained>
 void memcpy_call(benchmark::State& state) {
     static_assert(std::is_trivially_copyable_v<Contained>, "memcpy must only be called on trivially copyable types");
     const size_t r0      = static_cast<size_t>(state.range(0));
-    const auto in_buffer = random_vector<Contained>(r0);
-    std::vector<Contained> out_buffer(r0);
+    const auto in_buffer = random_vector<Contained, not_highly_aligned_allocator>(r0);
+    std::vector<Contained, not_highly_aligned_allocator<Contained>> out_buffer(r0);
     for ([[maybe_unused]] auto _ : state) {
         benchmark::DoNotOptimize(in_buffer.data());
         memcpy(out_buffer.data(), in_buffer.data(), r0 * sizeof(Contained));
@@ -62,8 +63,8 @@ void memcpy_call(benchmark::State& state) {
 template <typename Contained>
 void std_copy_call(benchmark::State& state) {
     const size_t r0      = static_cast<size_t>(state.range(0));
-    const auto in_buffer = random_vector<Contained>(r0);
-    std::vector<Contained> out_buffer(r0);
+    const auto in_buffer = random_vector<Contained, not_highly_aligned_allocator>(r0);
+    std::vector<Contained, not_highly_aligned_allocator<Contained>> out_buffer(r0);
     for ([[maybe_unused]] auto _ : state) {
         benchmark::DoNotOptimize(in_buffer.data());
         std::copy(in_buffer.begin(), in_buffer.end(), out_buffer.begin());
@@ -74,8 +75,8 @@ void std_copy_call(benchmark::State& state) {
 template <typename Contained>
 void std_copy_n_call(benchmark::State& state) {
     const size_t r0      = static_cast<size_t>(state.range(0));
-    const auto in_buffer = random_vector<Contained>(r0);
-    std::vector<Contained> out_buffer(r0);
+    const auto in_buffer = random_vector<Contained, not_highly_aligned_allocator>(r0);
+    std::vector<Contained, not_highly_aligned_allocator<Contained>> out_buffer(r0);
     for ([[maybe_unused]] auto _ : state) {
         benchmark::DoNotOptimize(in_buffer.data());
         std::copy_n(in_buffer.begin(), r0, out_buffer.begin());